Remove unused references to aerospace_post_sum (#73)
aerospace_post_sum is used to display "Page X/Y" at the top and bottom of the page. You can see it also used in the other archives (e.g. Aerospace 101). Because the data for the table is fetched in advanced, pagination through dataTables.js is not a true page refresh, so the aerospace_post_sum doesn't get triggered and "PAGE 1/2" does not update to "PAGE 2/2" when the user goes to the second page of the table.